Rename `replication_registry` module into `registry`
 Contains a lots of stuff and its already inside `replication` module.
