+++ src/shared.rs
@@ -13,8 +13,8 @@ use crate::prelude::*;
 use backend::connected_client::NetworkIdMap;
 use event::remote_event_registry::RemoteEventRegistry;
 use replication::{
-    command_markers::CommandMarkers, replication_registry::ReplicationRegistry,
-    rules::ReplicationRules, track_mutate_messages::TrackMutateMessages,
+    command_markers::CommandMarkers, registry::ReplicationRegistry, rules::ReplicationRules,
+    track_mutate_messages::TrackMutateMessages,
 };
 
 /// Initializes types, resources and events needed for both client and server.
