+++ src/shared/replication.rs
@@ -2,7 +2,7 @@ pub mod client_ticks;
 pub mod command_markers;
 pub mod deferred_entity;
 pub(crate) mod mutate_index;
-pub mod replication_registry;
+pub mod registry;
 pub mod rules;
 pub mod track_mutate_messages;
 pub mod update_message_flags;
