+++ src/shared/replication/registry.rs
