+++ src/shared/replication/registry/command_fns.rs
