+++ src/shared/replication/registry/component_fns.rs
