+++ src/shared/replication/registry/rule_fns.rs
